x264: use int64_t instead int as delay
(cherry-picked from commit e0600d460ae60e32b189b79ed61c9dc217577af2)

+++ modules/codec/x264.c
@@ -714,7 +714,7 @@ struct encoder_sys_t
     x264_param_t    param;
 
 #if X264_BUILD >= 83
-    int  i_initial_delay;
+    int64_t  i_initial_delay;
 #else
     mtime_t         i_interpolated_dts;
 #endif
